Fixed all warning on MSVC for IO

+++ source/io/TcpClient.cpp
@@ -254,7 +254,7 @@ bool TcpClient::is_open() const {
     return m_impl->is_open();
 }
 
-void TcpClient::send_data(std::shared_ptr<const char> buffer, std::size_t size, EndSendCallback callback) {
+void TcpClient::send_data(std::shared_ptr<const char> buffer, std::uint32_t size, EndSendCallback callback) {
     return m_impl->send_data(buffer, size, callback);
 }
