+++ source/io/TcpConnectedClient.h
@@ -31,7 +31,7 @@ public:
 
     IO_DLL_PUBLIC void set_close_callback(CloseCallback callback);
 
-    IO_DLL_PUBLIC void send_data(std::shared_ptr<const char> buffer, std::size_t size, EndSendCallback callback = nullptr);
+    IO_DLL_PUBLIC void send_data(std::shared_ptr<const char> buffer, std::uint32_t size, EndSendCallback callback = nullptr);
     IO_DLL_PUBLIC void send_data(const std::string& message, EndSendCallback callback = nullptr);
 
     // TODO: rename as pending_send_requesets??? Because name is inconsistent.
